cleanup rust code
